+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - API `/api/file/getFile` 支持自定义响应类型 | API `/api/file/getFile` supports custom response types
 - 添加 API `/api/inbox/getShorthand` | Add API `/api/inbox/getShorthand`
 - 添加 API `/api/history/getDocHistoryContent` | Add API `/api/history/getDocHistoryContent`
+- 添加 API `/api/history/getHistoryItems` | Add API `/api/history/getHistoryItems`
 
 ## v0.1.2 / 2023-07-26